flask-app/requirements.txt
#azure_storage
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.3.0
cryptography==38.0.4
Cython==0.29.33
dl==0.1.0
HTMLParser==0.0.2
importlib_metadata==4.11.3
ipaddr==2.2.0
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.1
mypy_extensions==1.0.0
numpy==1.21.6
ordereddict==1.1
pandas==1.3.5
protobuf==4.22.0
pyOpenSSL==23.0.0
python-dotenv==0.21.1
scipy==1.7.3
#wincertstore==0.2.1
#xmlrpclib==1.0.1
#zipp==3.11.0
requests
azure.storage.blob
azure-functions
flask
implicit
